Style sidenav in accordance with mockups
One deviation from the mockup: if a sub-item is selected, text weight remains
the same. This is so hovering over text does not cause text to wrap, shifting
the menu item's height in unexpected ways.

Mockup has other non-sidenav specific layout changes -- deferring those so they
can be addressed as a whole.